Issue #000 fix: QuestionPublisher implementation.

+++ publish-pipeline/publish-core/src/main/scala/org/sunbird/publish/helpers/ObjectReader.scala
@@ -7,14 +7,14 @@ import scala.collection.JavaConverters._
 
 trait ObjectReader {
 
-  def getObject(identifier: String, pkgVersion: Int)(implicit neo4JUtil: Neo4JUtil, cassandraUtil: CassandraUtil): ObjectData = {
+  def getObject(identifier: String, pkgVersion: Double)(implicit neo4JUtil: Neo4JUtil, cassandraUtil: CassandraUtil): ObjectData = {
     val metadata = getMetadata(identifier, pkgVersion)
     val extData = getExtData(identifier)
     val hierarchy = getHierarchy(identifier)
     ObjectData(identifier, metadata, extData, hierarchy)
   }
 
-  private def getMetadata(identifier: String, pkgVersion: Int)(implicit neo4JUtil: Neo4JUtil): Map[String, AnyRef] = {
+  private def getMetadata(identifier: String, pkgVersion: Double)(implicit neo4JUtil: Neo4JUtil): Map[String, AnyRef] = {
     val nodeId = if (pkgVersion > 0) identifier + ".img" else identifier
     neo4JUtil.getNodeProperties(nodeId).asScala.toMap
   }
